airflow-local/requirements.txt
apache-airflow[postgres]==2.0.0
sqlalchemy==1.3.19
psycopg2==2.8.6
